+++ example/indieauth.php
@@ -28,12 +28,10 @@ try {
     $service->setDefaultRoute('/welcome');
 
     // use discovery by default, fall back to IndieCert (https://indiecert.net/auth)
-    // redirect to '/success' after authentication
-    $indieAuth = new IndieAuthAuthentication('/success');
+    $indieAuth = new IndieAuthAuthentication();
 
     // use discovery by default, fall back to IndieAuth (https://indieauth.com/auth)
-    // redirect to '/success' after authentication
-    //$indieAuth = new IndieAuthAuthentication('/success', 'https://indieauth.com/auth');
+    //$indieAuth = new IndieAuthAuthentication('https://indieauth.com/auth');
 
     // disable discovery (i.e. "Distributed IndieAuth")
     // $indieAuth->setDiscovery(false);
@@ -44,7 +42,7 @@ try {
         '/welcome',
         function (Request $request) {
             // Show Sign In form;  POST to 'indieauth/auth' endpoint which is registered by the IndieAuth plugin
-            return '<html><head></head><body><h1>Sign In</h1><form method="post" action="indieauth/auth">https://<input type="text" name="me" placeholder="example.org" required><input type="submit" value="Sign In"></form></body></html>';
+            return '<html><head></head><body><h1>Sign In</h1><form method="post" action="indieauth/auth">https://<input type="text" name="me" placeholder="example.org" required><input type="hidden" name="redirect_to" value="/success"><input type="submit" value="Sign In"></form></body></html>';
         },
         array(
             // To view the "welcome" page, no authentication is required